requirements.txt
# Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Utilities
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
